Stepless scrolling workaround to mouse wheel event
Stepless scrolling is present when using trackpad or mouse that has stepless scroll weel (not in Firefox, because of different event).
This workaround fixes mouse wheel event in mxEvent and a zooming bug in Graph Editor relating this issue.
Target and custom filter had to be implemented to the event, because while zooming in Graph Editor with stepless mouse you need to prevent the default behavior of the event, otherwise scrolling would happen if the movement was too small.

Tested using Google Chrome version 60.0 with trackpad.
Further testing with different browsers, mouses and examples adviced.